% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{dice}
\alias{dice}
\title{Sorensen-Dice coefficient}
\usage{
dice(size_a, size_b, intersection)
}
\arguments{
\item{size_a}{Inclusive size of set A (integer >= 0).}

\item{size_b}{Inclusive size of set B (integer >= 0).}

\item{intersection}{Inclusive intersection size |A intersection B|.}
}
\value{
Numeric in [0, 1].
}
\description{
Computes 2 * |A intersection B| / (|A| + |B|). Returns 0 if both sets are
empty (matches web tool convention).
}
\examples{
dice(10, 10, 5)
}
