% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{list_samples}
\alias{list_samples}
\title{List bundled sample dataset names}
\usage{
list_samples()
}
\value{
Character vector of 5 sample identifiers.
}
\description{
Returns the names of the 5 bundled sample datasets, sorted alphabetically.
Use [load_sample()] to load one.
}
\examples{
list_samples()
}
