% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{overlap_coefficient}
\alias{overlap_coefficient}
\title{Szymkiewicz-Simpson overlap coefficient}
\usage{
overlap_coefficient(size_a, size_b, intersection)
}
\arguments{
\item{size_a}{Inclusive size of set A (integer >= 0).}

\item{size_b}{Inclusive size of set B (integer >= 0).}

\item{intersection}{Inclusive intersection size |A intersection B|.}
}
\value{
Numeric in [0, 1].
}
\description{
Computes |A intersection B| / min(|A|, |B|). Useful when one set is much
smaller than the other.
}
\examples{
overlap_coefficient(10, 5, 3)
}
